requirements.txt
deepl==1.8.0
pytesseract==0.3.9
Pillow==9.2.0
Wand==0.6.7
streamlit==1.10.0
